requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt --extra dev
aiosqlite==0.21.0
    # via fastapi-template (pyproject.toml)
alembic==1.16.1
    # via fastapi-template (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
asyncpg==0.30.0
    # via fastapi-template (pyproject.toml)
bcrypt==4.3.0
    # via passlib
black==25.1.0
    # via fastapi-template (pyproject.toml)
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
click==8.2.1
    # via
    #   black
    #   uvicorn
coverage==7.9.1
    # via
    #   fastapi-template (pyproject.toml)
    #   pytest-cov
cryptography==45.0.2
    # via python-jose
distlib==0.3.9
    # via virtualenv
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via fastapi-template (pyproject.toml)
fastapi==0.115.12
    # via fastapi-template (pyproject.toml)
filelock==3.18.0
    # via virtualenv
greenlet==3.2.2
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.27.2
    # via fastapi-template (pyproject.toml)
identify==2.6.12
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via fastapi-template (pyproject.toml)
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
mypy==1.16.1
    # via fastapi-template (pyproject.toml)
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   black
    #   pytest
passlib==1.7.4
    # via fastapi-template (pyproject.toml)
pathspec==0.12.1
    # via
    #   black
    #   mypy
platformdirs==4.3.8
    # via
    #   black
    #   virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.2.0
    # via fastapi-template (pyproject.toml)
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   fastapi-template (pyproject.toml)
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via fastapi-template (pyproject.toml)
pygments==2.19.1
    # via pytest
pytest==8.4.0
    # via
    #   fastapi-template (pyproject.toml)
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==1.0.0
    # via fastapi-template (pyproject.toml)
pytest-cov==6.2.1
    # via fastapi-template (pyproject.toml)
python-dotenv==1.1.0
    # via
    #   fastapi-template (pyproject.toml)
    #   pydantic-settings
python-jose==3.4.0
    # via fastapi-template (pyproject.toml)
python-multipart==0.0.20
    # via fastapi-template (pyproject.toml)
pyyaml==6.0.2
    # via pre-commit
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via
    #   fastapi-template (pyproject.toml)
    #   alembic
starlette==0.46.2
    # via fastapi
typing-extensions==4.13.2
    # via
    #   aiosqlite
    #   alembic
    #   anyio
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.34.2
    # via fastapi-template (pyproject.toml)
virtualenv==20.31.2
    # via pre-commit
openai==1.51.2
pydub==0.25.1
requests==2.32.3
markdown-it-py==3.0.0
reportlab==4.2.2